Conflicts:
	component/admin/elements/orderbyproduct.php
	component/admin/extras/sh404sef/sef_ext/com_redshop.php
	component/admin/helpers/product_category.php
	component/site/helpers/extrafield.php
	component/site/helpers/producthelper.php
	component/site/helpers/redhelper.php
	component/site/helpers/rscarthelper.php
	component/site/helpers/rsuserhelper.php
	modules/site/mod_redmanufacturer/mod_redmanufacturer.php
	modules/site/mod_redproducttab/mod_redproducttab.php
	modules/site/mod_redproducttab/tmpl/default.php
	plugins/redshop_payment/rs_payment_ewayuk/rs_payment_ewayuk/eway_response.php

+++ component/admin/helpers/redshopmail.php
@@ -1318,17 +1318,17 @@ class redshopMail
 		{
 			$tax = $row->quotation_tax;
 
-			if (VAT_RATE_AFTER_DISCOUNT)
+			if ((float) VAT_RATE_AFTER_DISCOUNT)
 			{
-				$Discountvat             = (VAT_RATE_AFTER_DISCOUNT * $row->quotation_discount) / (1 + VAT_RATE_AFTER_DISCOUNT);
+				$Discountvat             = ((float) VAT_RATE_AFTER_DISCOUNT * $row->quotation_discount) / (1 + (float) VAT_RATE_AFTER_DISCOUNT);
 				$row->quotation_discount = $row->quotation_discount - $Discountvat;
 				$tax                     = $tax - $Discountvat;
 			}
 
-			if (VAT_RATE_AFTER_DISCOUNT)
+			if ((float) VAT_RATE_AFTER_DISCOUNT)
 			{
 				$sp_discount             = ($row->quotation_special_discount * ($row->quotation_subtotal + $row->quotation_tax)) / 100;
-				$Discountspvat           = ($sp_discount * VAT_RATE_AFTER_DISCOUNT) / (1 + VAT_RATE_AFTER_DISCOUNT);
+				$Discountspvat           = ($sp_discount * (float) VAT_RATE_AFTER_DISCOUNT) / (1 + (float) VAT_RATE_AFTER_DISCOUNT);
 				$DiscountspWithotVat     = $sp_discount - $Discountspvat;
 				$row->quotation_discount = $row->quotation_discount + $DiscountspWithotVat;
 				$tax                     = $tax - $Discountspvat;
